+++ composer.json
@@ -18,6 +18,7 @@
   "require-dev": {
     "codeception/codeception": "^2.2",
     "codeception/aspect-mock": "*",
+    "phpro/grumphp": "^0.14.2",
     "squizlabs/php_codesniffer": "3.*"
   },
   "autoload": {
