+++ grumphp.yml
@@ -1,0 +1,20 @@
+parameters:
+    git_dir: .
+    bin_dir: vendor/bin
+    ascii:
+        failed: ~
+        succeeded: ~
+    tasks:
+        phplint:
+            triggered_by: ['php', 'inc', 'phtml']
+        phpcs:
+            encoding: UTF-8
+            triggered_by: ['php', 'inc', 'phtml']
+            whitelist_patterns:
+                - /^src\/(.*)/
+                - /^tests\/unit\/(.*)/
+        codeception:
+            config_file: ~
+            fail_fast: false
+            suite: unit
+            test: ~
