+++ tests/unit/MotherObject/TestClient.php
@@ -1,0 +1,9 @@
+<?php
+namespace Tests\Fei\ApiClient\MotherObject;
+
+use Fei\ApiClient\AbstractApiClient;
+
+class TestClient extends AbstractApiClient
+{
+    const OPTION_TEST = 'testOption';
+}
